# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OmniGest",
        "description": "Documentação da API Omnigest",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://sistema-generico-api.vercel.app/api/api"
        }
    ],
    "paths": {
...
        "/estoque/{id_estoque}": {
            "get": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Obtém um estoque pelo ID",
                "operationId": "276a5b0c1316c3ba9337fa2ad35fcb43",
                "parameters": [
                    {
                        "name": "id_estoque",
                        "in": "path",
                        "description": "ID do estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estoque encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Estoque"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Estoque não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Atualiza um estoque",
                "operationId": "dc56ceaae9eb25a035c2bfa4ed8aa4f7",
                "parameters": [
                    {
                        "name": "id_estoque",
                        "in": "path",
                        "description": "ID do estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Estoque"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estoque atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Deleta um estoque",
                "operationId": "6285ba21235806c605c51620eb5c0c7e",
                "parameters": [
                    {
                        "name": "id_estoque",
                        "in": "path",
                        "description": "ID do estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estoque deletado com sucesso"
                    }
                }
            }
        },
        "/estoque": {
            "post": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Cria um novo estoque",
                "operationId": "8deb717d7f84593ec996a1e5927ad48a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Estoque"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estoque criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Estoque"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Estoque": {
                "properties": {
                    "id_estoque_est": {
                        "type": "integer"
                    },
                    "id_empresa_est": {
                        "description": "Chave estrangeira referenciando a empresa",
                        "type": "integer"
                    },
                    "des_estoque_est": {
                        "type": "string"
                    },
                    "id_centro_custo_est": {
                        "description": "Chave estrangeira referenciando o centro de custo",
                        "type": "integer"
                    },
                    "is_ativo_est": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}